v 5.85 - SmartConfig using esp-touch
-New Features:
You can now configure the wifi settings on startup using the esptouch
app available in the app store / google play.  Note that this app does
not appear to work yet with ios 12.  You will have 60 seconds to use the
app to configure it, then it will go into the default AP setup mode.